Drop the main-module guard from the release check CLI
A misfiring guard would let pnpm release:check exit 0 without checking
anything. The CLI now always runs; the checks live in release-check.ts.
